requirements.txt
pdfplumber
pypdf
pillow
pytesseract
pandas
openpyxl
python-dateutil
pydantic
rapidfuzz
ruamel.yaml
typer
loguru
